fix int mm

+++ src/quant_mp/QModules.py
@@ -63,7 +63,7 @@ class QLinearFunction(Function):
             weight = qweight.dequant(weight, (scale_bw[0], 0.))
             scale_bw[0] = torch.ones_like(qweight.s)
 
-        output = scale_bw[1] * linear(input, weight, None) * scale_bw[0].T
+        output = scale_bw[1] * linear(input.float(), weight.float(), None) * scale_bw[0].T
         if bias is not None:
             output += bias.unsqueeze(0).expand_as(output)
